[Synthetics] Rename inspect fixture to avoid scout lint false-positive
The @kbn/eslint/scout_test_file_naming rule flags non-tests/ scout files
whose basename contains the substring "spec". "inspect" matched it, so
the data fixture inspect_browser_monitor.ts tripped the Linting step.
Rename it to browser_monitor_fixture.ts (it is a fixture, not a test).
