+++ x-pack/solutions/observability/plugins/synthetics/test/scout/api/fixtures/data/browser_monitor_fixture.ts
